+++ frontend/app/src/entities/events/ui/filters/filter-kind-select.tsx
