+++ frontend/app/src/entities/events/ui/filters/global-branch-filter.tsx
