+++ frontend/app/src/entities/events/ui/filters/global-filter-form.tsx
