+++ frontend/app/src/entities/events/ui/filters/global-kind-filter-form.tsx
